非洲黑猩猩3994

+++ application/IDEA/app/source/Control/GenerationResult/GameResult_3998.c
@@ -255,7 +255,7 @@ void GameResult_3998_GenBonus(RoundInfo_t* info,int32_t betVal,GameInstance_t* i
 			uint8_t BlindBoxCountArray[3] = { 6, 8, 11 };
 			uint8_t BlindBoxCount = BlindBoxCountArray[bonusCount - 3];
 
-			int dataArray[15] = { 0, 1, 2, 3 , 4 ,5,6,7,8,9,10,11,12,13,14 };
+			uint8_t dataArray[15] = { 0, 1, 2, 3 , 4 ,5,6,7,8,9,10,11,12,13,14 };
 			uint8_t  PosVec[15];
 			uint8_t  PosSize = 15;
